refactor(stm): rename ivc halo2 snark
