+++ mithril-stm/src/proof_system/ivc_halo2_snark/prover_input.rs
