+++ mithril-stm/src/proof_system/ivc_halo2_snark/rolling_state.rs
